pyproject.toml
[tool.poetry]
name = "iguana"
version = "0.1.0"
description = ""
authors = ["Jorge Vilaça"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2"
psycopg2-binary = "^2.8.6"
djangorestframework = "^3.12.4"
django-localflavor = "^3.0.1"

[tool.poetry.dev-dependencies]
black = {version = "^21.4b2", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
